pyapp/requirements.txt
google-cloud-storage
google-auth